Immutability is the notion that something cannot be changed.
It has become a buzzword in the developer community lately because of its pervasiveness throughout modern languages like ClosureScript and Elm. Why would you want data that can never change, when data is constantly changing in modern applications? One of the harder problems to solve for is how to detect when data has been updated. Once you know that, you can update all of the pieces that need to respond.
Let’s look at a common case in JavaScript where we want to check if a value has changed.

There’s no trick to this example. We’re taking a variable named data, and creating a newData variable that is equal to the original. When we check if the two variables are equal, it’s true as expected. When we change the value of newData and check the equality again, we see that they are unequal, thus the newData has changed.
This works because when we declare the data variable, it is immediately given a value. Then we do the same with newData and set its value to the value of data. When we set the value of newData to something else, it is not equal to the value stored in the data variable.
So far, there’s no problem, right? With primitives in JavaScript, when we declare a variable, that variable effectively is a placeholder for its value. When we change the value, the variable changes. Taking that a step further, changing its value changes the variable itself.
This isn’t true for non-primitives in JavaScript. Let’s see how this breaks down with objects.

Thanks to
[this](https://medium.com/@brooklynunited/the-case-for-immutability-d4e45029b8bc)
blog for helping me write this PR.